+++ Appium Wizard/MainScreen.Designer.cs
@@ -55,6 +55,7 @@ namespace Appium_Wizard
             fixInstallationToolStripMenuItem = new ToolStripMenuItem();
             updaterToolStripMenuItem = new ToolStripMenuItem();
             pluginsToolStripMenuItem1 = new ToolStripMenuItem();
+            restartADBServerToolStripMenuItem = new ToolStripMenuItem();
             toolsToolStripMenuItem = new ToolStripMenuItem();
             inspectorToolStripMenuItem1 = new ToolStripMenuItem();
             iOSProfileManagementToolStripMenuItem = new ToolStripMenuItem();
@@ -127,6 +128,7 @@ namespace Appium_Wizard
             OpenDropDownButton = new Button();
             openContextMenuStrip = new ContextMenuStrip(components);
             reInitializeDeviceToolStripMenuItem = new ToolStripMenuItem();
+            unlockDeviceToolStripMenuItem = new ToolStripMenuItem();
             contextMenuStrip1.SuspendLayout();
             menuStrip1.SuspendLayout();
             tabControl1.SuspendLayout();
@@ -241,6 +243,7 @@ namespace Appium_Wizard
             // 
             // androidWiFiToolStripMenuItem
             // 
+            androidWiFiToolStripMenuItem.Image = Properties.Resources.wifi;
             androidWiFiToolStripMenuItem.Name = "androidWiFiToolStripMenuItem";
             androidWiFiToolStripMenuItem.Size = new Size(156, 30);
             androidWiFiToolStripMenuItem.Text = "Android Wi-Fi";
@@ -300,7 +303,7 @@ namespace Appium_Wizard
             // 
             // fileToolStripMenuItem
             // 
-            fileToolStripMenuItem.DropDownItems.AddRange(new ToolStripItem[] { serverSetupToolStripMenuItem, fixInstallationToolStripMenuItem, updaterToolStripMenuItem, pluginsToolStripMenuItem1 });
+            fileToolStripMenuItem.DropDownItems.AddRange(new ToolStripItem[] { serverSetupToolStripMenuItem, fixInstallationToolStripMenuItem, updaterToolStripMenuItem, pluginsToolStripMenuItem1, restartADBServerToolStripMenuItem });
             fileToolStripMenuItem.Name = "fileToolStripMenuItem";
             fileToolStripMenuItem.Size = new Size(51, 20);
             fileToolStripMenuItem.Text = "Server";
@@ -309,7 +312,7 @@ namespace Appium_Wizard
             // 
             serverSetupToolStripMenuItem.Image = Properties.Resources.gear;
             serverSetupToolStripMenuItem.Name = "serverSetupToolStripMenuItem";
-            serverSetupToolStripMenuItem.Size = new Size(163, 22);
+            serverSetupToolStripMenuItem.Size = new Size(171, 22);
             serverSetupToolStripMenuItem.Text = "Configuration";
             serverSetupToolStripMenuItem.Click += serverSetupToolStripMenuItem_Click;
             // 
@@ -317,7 +320,7 @@ namespace Appium_Wizard
             // 
             fixInstallationToolStripMenuItem.Image = Properties.Resources.troubleshooting;
             fixInstallationToolStripMenuItem.Name = "fixInstallationToolStripMenuItem";
-            fixInstallationToolStripMenuItem.Size = new Size(163, 22);
+            fixInstallationToolStripMenuItem.Size = new Size(171, 22);
             fixInstallationToolStripMenuItem.Text = "Troubleshooter";
             fixInstallationToolStripMenuItem.Click += fixInstallationToolStripMenuItem_Click;
             // 
@@ -325,7 +328,7 @@ namespace Appium_Wizard
             // 
             updaterToolStripMenuItem.Image = Properties.Resources.update;
             updaterToolStripMenuItem.Name = "updaterToolStripMenuItem";
-            updaterToolStripMenuItem.Size = new Size(163, 22);
+            updaterToolStripMenuItem.Size = new Size(171, 22);
             updaterToolStripMenuItem.Text = "Updater";
             updaterToolStripMenuItem.Click += updaterToolStripMenuItem_Click;
             // 
@@ -333,9 +336,17 @@ namespace Appium_Wizard
             // 
             pluginsToolStripMenuItem1.Image = Properties.Resources.plug_in;
             pluginsToolStripMenuItem1.Name = "pluginsToolStripMenuItem1";
-            pluginsToolStripMenuItem1.Size = new Size(163, 22);
+            pluginsToolStripMenuItem1.Size = new Size(171, 22);
             pluginsToolStripMenuItem1.Text = "Plugins Manager";
             pluginsToolStripMenuItem1.Click += pluginsToolStripMenuItem1_Click;
+            // 
+            // restartADBServerToolStripMenuItem
+            // 
+            restartADBServerToolStripMenuItem.Image = Properties.Resources.adb;
+            restartADBServerToolStripMenuItem.Name = "restartADBServerToolStripMenuItem";
+            restartADBServerToolStripMenuItem.Size = new Size(171, 22);
+            restartADBServerToolStripMenuItem.Text = "Restart ADB Server";
+            restartADBServerToolStripMenuItem.Click += restartADBServerToolStripMenuItem_Click;
             // 
             // toolsToolStripMenuItem
             // 
@@ -442,7 +453,7 @@ namespace Appium_Wizard
             appiumDocsToolStripMenuItem.DropDownItems.AddRange(new ToolStripItem[] { capabilitiesToolStripMenuItem, settingsAPIToolStripMenuItem, pluginsToolStripMenuItem, xCUITestToolStripMenuItem, uIAutomatorToolStripMenuItem, serverSecurityToolStripMenuItem, cLIArgumentsToolStripMenuItem });
             appiumDocsToolStripMenuItem.Image = Properties.Resources.doc;
             appiumDocsToolStripMenuItem.Name = "appiumDocsToolStripMenuItem";
-            appiumDocsToolStripMenuItem.Size = new Size(255, 30);
+            appiumDocsToolStripMenuItem.Size = new Size(247, 22);
             appiumDocsToolStripMenuItem.Text = "Appium Docs";
             // 
             // capabilitiesToolStripMenuItem
@@ -554,7 +565,7 @@ namespace Appium_Wizard
             otherDocsToolStripMenuItem.DropDownItems.AddRange(new ToolStripItem[] { iOSNativeAppsBundleToolStripMenuItem });
             otherDocsToolStripMenuItem.Image = Properties.Resources.doc;
             otherDocsToolStripMenuItem.Name = "otherDocsToolStripMenuItem";
-            otherDocsToolStripMenuItem.Size = new Size(255, 30);
+            otherDocsToolStripMenuItem.Size = new Size(247, 22);
             otherDocsToolStripMenuItem.Text = "Other Docs";
             // 
             // iOSNativeAppsBundleToolStripMenuItem
@@ -569,7 +580,7 @@ namespace Appium_Wizard
             // 
             troubleshootToolStripMenuItem.Image = Properties.Resources.guide;
             troubleshootToolStripMenuItem.Name = "troubleshootToolStripMenuItem";
-            troubleshootToolStripMenuItem.Size = new Size(255, 30);
+            troubleshootToolStripMenuItem.Size = new Size(247, 22);
             troubleshootToolStripMenuItem.Text = "Troubleshooting Guide";
             troubleshootToolStripMenuItem.Click += fAQToolStripMenuItem_Click;
             // 
@@ -577,7 +588,7 @@ namespace Appium_Wizard
             // 
             reportAnIssueToolStripMenuItem.Image = Properties.Resources.bug;
             reportAnIssueToolStripMenuItem.Name = "reportAnIssueToolStripMenuItem";
-            reportAnIssueToolStripMenuItem.Size = new Size(255, 30);
+            reportAnIssueToolStripMenuItem.Size = new Size(247, 22);
             reportAnIssueToolStripMenuItem.Text = "Report an Issue | Feature Request";
             reportAnIssueToolStripMenuItem.Click += reportAnIssueToolStripMenuItem_Click;
             // 
@@ -585,7 +596,7 @@ namespace Appium_Wizard
             // 
             openLogsFolderToolstripMenuItem.Image = Properties.Resources.log;
             openLogsFolderToolstripMenuItem.Name = "openLogsFolderToolstripMenuItem";
-            openLogsFolderToolstripMenuItem.Size = new Size(255, 30);
+            openLogsFolderToolstripMenuItem.Size = new Size(247, 22);
             openLogsFolderToolstripMenuItem.Text = "Open Logs folder";
             openLogsFolderToolstripMenuItem.Click += openLogsFolderToolstripMenuItem_Click;
             // 
@@ -593,7 +604,7 @@ namespace Appium_Wizard
             // 
             startADiscussionToolStripMenuItem.Image = Properties.Resources.Discussion;
             startADiscussionToolStripMenuItem.Name = "startADiscussionToolStripMenuItem";
-            startADiscussionToolStripMenuItem.Size = new Size(255, 30);
+            startADiscussionToolStripMenuItem.Size = new Size(247, 22);
             startADiscussionToolStripMenuItem.Text = "Start a Discussion";
             startADiscussionToolStripMenuItem.Click += startADiscussionToolStripMenuItem_Click;
             // 
@@ -601,7 +612,7 @@ namespace Appium_Wizard
             // 
             checkForUpdatesToolStripMenuItem.Image = Properties.Resources.update;
             checkForUpdatesToolStripMenuItem.Name = "checkForUpdatesToolStripMenuItem";
-            checkForUpdatesToolStripMenuItem.Size = new Size(255, 30);
+            checkForUpdatesToolStripMenuItem.Size = new Size(247, 22);
             checkForUpdatesToolStripMenuItem.Text = "Check for Updates...";
             checkForUpdatesToolStripMenuItem.Click += checkForUpdatesToolStripMenuItem_Click;
             // 
@@ -609,7 +620,7 @@ namespace Appium_Wizard
             // 
             changeLogToolStripMenuItem.Image = Properties.Resources.changelog;
             changeLogToolStripMenuItem.Name = "changeLogToolStripMenuItem";
-            changeLogToolStripMenuItem.Size = new Size(255, 30);
+            changeLogToolStripMenuItem.Size = new Size(247, 22);
             changeLogToolStripMenuItem.Text = "Changelog";
             changeLogToolStripMenuItem.Click += changeLogToolStripMenuItem_Click;
             // 
@@ -617,7 +628,7 @@ namespace Appium_Wizard
             // 
             readMeToolStripMenuItem.Image = Properties.Resources.readme;
             readMeToolStripMenuItem.Name = "readMeToolStripMenuItem";
-            readMeToolStripMenuItem.Size = new Size(255, 30);
+            readMeToolStripMenuItem.Size = new Size(247, 22);
             readMeToolStripMenuItem.Text = "Read me";
             readMeToolStripMenuItem.Click += readMeToolStripMenuItem_Click;
             // 
@@ -625,7 +636,7 @@ namespace Appium_Wizard
             // 
             aboutToolStripMenuItem.Image = Properties.Resources.information;
             aboutToolStripMenuItem.Name = "aboutToolStripMenuItem";
-            aboutToolStripMenuItem.Size = new Size(255, 30);
+            aboutToolStripMenuItem.Size = new Size(247, 22);
             aboutToolStripMenuItem.Text = "About";
             aboutToolStripMenuItem.Click += aboutToolStripMenuItem_Click;
             // 
@@ -849,9 +860,9 @@ namespace Appium_Wizard
             // contextMenuStrip4
             // 
             contextMenuStrip4.ImageScalingSize = new Size(24, 24);
-            contextMenuStrip4.Items.AddRange(new ToolStripItem[] { installAppToolStripMenuItem, launchAppToolStripMenuItem, refreshStatusToolStripMenuItem, takeScreenshotToolStripMenuItem, rebootDeviceToolStripMenuItem });
+            contextMenuStrip4.Items.AddRange(new ToolStripItem[] { installAppToolStripMenuItem, launchAppToolStripMenuItem, refreshStatusToolStripMenuItem, takeScreenshotToolStripMenuItem, unlockDeviceToolStripMenuItem, rebootDeviceToolStripMenuItem });
             contextMenuStrip4.Name = "contextMenuStrip4";
-            contextMenuStrip4.Size = new Size(240, 154);
+            contextMenuStrip4.Size = new Size(240, 206);
             // 
             // installAppToolStripMenuItem
             // 
@@ -986,6 +997,14 @@ namespace Appium_Wizard
             reInitializeDeviceToolStripMenuItem.Size = new Size(173, 22);
             reInitializeDeviceToolStripMenuItem.Text = "Re-Initialize Device";
             reInitializeDeviceToolStripMenuItem.Click += reInitializeDeviceToolStripMenuItem_Click;
+            // 
+            // unlockDeviceToolStripMenuItem
+            // 
+            unlockDeviceToolStripMenuItem.Image = Properties.Resources.Unlock;
+            unlockDeviceToolStripMenuItem.Name = "unlockDeviceToolStripMenuItem";
+            unlockDeviceToolStripMenuItem.Size = new Size(239, 30);
+            unlockDeviceToolStripMenuItem.Text = "Unlock Device";
+            unlockDeviceToolStripMenuItem.Click += unlockDeviceToolStripMenuItem_Click;
             // 
             // MainScreen
             // 
@@ -1137,5 +1156,7 @@ namespace Appium_Wizard
         private ToolStripMenuItem reInitializeDeviceToolStripMenuItem;
         private ToolStripMenuItem changeLogToolStripMenuItem;
         private ToolStripMenuItem readMeToolStripMenuItem;
+        private ToolStripMenuItem restartADBServerToolStripMenuItem;
+        private ToolStripMenuItem unlockDeviceToolStripMenuItem;
     }
 }