+++ Appium Wizard/ScreenControl.cs
@@ -124,6 +124,7 @@ namespace Appium_Wizard
                 deviceSerialNumber = iOSMethods.GetInstance().GetDeviceSerialNumber(udid);
             }
             Logger.Info("Initialization completed");
+            this.KeyPreview = true;
         }
 
         public void UpdateStatusLabel(ScreenControl screenControl, string actualText)
@@ -367,7 +368,6 @@ namespace Appium_Wizard
             }
         }
 
-
         public async void WebView_MouseUp(object sender, MouseEventArgs e)
         {
             moveToX = e.Location.X;
@@ -383,6 +383,7 @@ namespace Appium_Wizard
                 // Check if the movement qualifies as a swipe
                 if (Math.Abs(deltaX) < swipeThreshold && Math.Abs(deltaY) < swipeThreshold)
                 {
+                    // Handle tap
                     if (isAndroid)
                     {
                         await Task.Run(() =>
@@ -402,90 +403,24 @@ namespace Appium_Wizard
                     return;
                 }
 
-                // Calculate screen-based swipe coordinates
-                int horizontalSwipeStartX = (int)(width * 0.1); // Start swipe at 10% of screen width
-                int horizontalSwipeEndX = (int)(width * 0.9);   // End swipe at 90% of screen width
-                int verticalSwipeStartY = (int)(height * 0.5);  // Start swipe at 50% of screen height
-                int verticalSwipeEndY = (int)(height * 0.9);    // End swipe at 90% of screen height
-                if (Math.Abs(deltaX) > Math.Abs(deltaY))
+                // Use actual mouse coordinates for swipe
+                int waitDuration = 300;
+
+                if (isAndroid)
                 {
-                    int waitDuration = 100;
-                    if (deltaX > 0)
-                    {   // Swipe right
-                        if (isAndroid)
-                        {
-                            await Task.Run(() =>
-                            {
-                                AndroidMethods.GetInstance().Swipe(udid, horizontalSwipeStartX, height / 2, horizontalSwipeEndX, height / 2, waitDuration);
-                            });
-                        }
-                        else
-                        {
-                            await Task.Run(() =>
-                            {
-                                iOSAPIMethods.Swipe(URL, sessionId, horizontalSwipeStartX, height / 2, horizontalSwipeEndX, height / 2, waitDuration);
-                            });
-                        }
-                    }
-                    else
+                    await Task.Run(() =>
                     {
-                        // Swipe left
-                        if (isAndroid)
-                        {
-                            await Task.Run(() =>
-                            {
-                                AndroidMethods.GetInstance().Swipe(udid, horizontalSwipeEndX, height / 2, horizontalSwipeStartX, height / 2, waitDuration);
-                            });
-                        }
-                        else
-                        {
-                            await Task.Run(() =>
-                            {
-                                iOSAPIMethods.Swipe(URL, sessionId, horizontalSwipeEndX, height / 2, horizontalSwipeStartX, height / 2, waitDuration);
-                            });
-                        }
-                    }
+                        AndroidMethods.GetInstance().SwipeForScreenControl(udid, pressX, pressY, moveToX, moveToY, waitDuration);
+                    });
                 }
                 else
                 {
-                    int waitDuration = 1000;
-                    if (deltaY > 0)
+                    await Task.Run(() =>
                     {
-                        // Swipe down
-                        if (isAndroid)
-                        {
-                            await Task.Run(() =>
-                            {
-                                AndroidMethods.GetInstance().Swipe(udid, width / 2, verticalSwipeStartY, width / 2, verticalSwipeEndY, waitDuration);
-                            });
-                        }
-                        else
-                        {
-                            await Task.Run(() =>
-                            {
-                                iOSAPIMethods.Swipe(URL, sessionId, width / 2, verticalSwipeStartY, width / 2, verticalSwipeEndY, waitDuration);
-                            });
-                        }
-                    }
-                    else
-                    {
-                        // Swipe up
-                        if (isAndroid)
-                        {
-                            await Task.Run(() =>
-                            {
-                                AndroidMethods.GetInstance().Swipe(udid, width / 2, verticalSwipeEndY, width / 2, verticalSwipeStartY, waitDuration);
-                            });
-                        }
-                        else
-                        {
-                            await Task.Run(() =>
-                            {
-                                iOSAPIMethods.Swipe(URL, sessionId, width / 2, verticalSwipeEndY, width / 2, verticalSwipeStartY, waitDuration);
-                            });
-                        }
-                    }
+                        iOSAPIMethods.Swipe(URL, sessionId, pressX, pressY, moveToX, moveToY, waitDuration);
+                    });
                 }
+
                 if (isAndroid)
                 {
                     GoogleAnalytics.SendEvent("SwipeScreen", "Android");
@@ -1260,7 +1195,7 @@ namespace Appium_Wizard
                     ActionType = "Set Device",
                 });
                 MessageBox.Show("Recording steps started.", "Record Steps", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                GoogleAnalytics.SendEvent("Recording steps started",OSType);
+                GoogleAnalytics.SendEvent("Recording steps started", OSType);
             }
             else
             {
@@ -1359,6 +1294,68 @@ namespace Appium_Wizard
         private void copyUDIDToolStripMenuItem_Click(object sender, EventArgs e)
         {
             Clipboard.SetText(this.udid);
+        }
+
+        private void ScreenWebView_KeyDown(object sender, KeyEventArgs e)
+        {
+            if (e.KeyCode == Keys.Back)
+            {
+                if (OSType.Equals("iOS"))
+                {
+                    iOSAPIMethods.BackSpace(URL, sessionId);
+                }
+                else
+                {
+                    AndroidMethods.GetInstance().BackSpace(udid);
+                }
+            }
+            else if (e.KeyCode == Keys.Delete)
+            {
+                if (OSType.Equals("Android"))
+                {
+                    AndroidMethods.GetInstance().Delete(udid);
+                }
+                //iOS don't have delete option
+            }
+            else if (e.KeyCode == Keys.Up)
+            {
+                if (OSType.Equals("Android"))
+                {
+                    AndroidMethods.GetInstance().UpArrow(udid);
+                }
+            }
+            else if (e.KeyCode == Keys.Down)
+            {
+                if (OSType.Equals("Android"))
+                {
+                    AndroidMethods.GetInstance().DownArrow(udid);
+                }
+            }
+            else if (e.KeyCode == Keys.Left)
+            {
+                if (OSType.Equals("Android"))
+                {
+                    AndroidMethods.GetInstance().LeftArrow(udid);
+                }
+            }
+            else if (e.KeyCode == Keys.Right)
+            {
+                if (OSType.Equals("Android"))
+                {
+                    AndroidMethods.GetInstance().RightArrow(udid);
+                }
+            }
+        }
+
+        private void ScreenWebView_PreviewKeyDown(object sender, PreviewKeyDownEventArgs e)
+        {
+            Debug.WriteLine($"PreviewKeyDown: {e.KeyCode}");
+            if (e.KeyCode == Keys.Back || e.KeyCode == Keys.Delete ||
+                e.KeyCode == Keys.Up || e.KeyCode == Keys.Down ||
+                e.KeyCode == Keys.Left || e.KeyCode == Keys.Right)
+            {
+                e.IsInputKey = true;
+            }
         }
     }